Time: 36 ms (96.58%), Space: 165.7 MB (5.08%) - LeetHub

+++ 2583-kth-largest-sum-in-a-binary-tree/2583-kth-largest-sum-in-a-binary-tree.cpp
@@ -17,7 +17,7 @@ public:
         q.push(root);
         while(!q.empty()){
             int size = q.size();
-            vector<long long> level;
+            vector<int> level;
             for(int i=0; i<size; i++){
                 TreeNode *node = q.front();
                 q.pop();
